pusher_center/requirements.txt
asgiref==3.3.1
autopep8==1.5.4
botnoi==0.2.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-cors-headers==3.6.0
django-filter==2.4.0
django-request==1.5.6
djangorestframework==3.12.2
idna==2.10
importlib-metadata==3.3.0
joblib==1.0.0
Markdown==3.3.3
numpy==1.19.5
Pillow==8.1.0
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
scikit-learn==0.24.1
scipy==1.5.4
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
zipp==3.4.0
pusher==3.0.0
esdk-obs-python==3.21.8
pymysql==1.0.2
mysqlclient==2.0.3
